+++ methods.go
@@ -943,3 +943,129 @@ func (b *Bot) RemoveChatVerification(ctx context.Context, params *RemoveChatVeri
 	err := b.rawRequest(ctx, "removeChatVerification", params, &result)
 	return result, err
 }
+
+// ReadBusinessMessage https://core.telegram.org/bots/api#readbusinessmessage
+func (b *Bot) ReadBusinessMessage(ctx context.Context, params *ReadBusinessMessageParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "readBusinessMessage", params, &result)
+	return result, err
+}
+
+// DeleteBusinessMessages https://core.telegram.org/bots/api#deletebusinessmessages
+func (b *Bot) DeleteBusinessMessages(ctx context.Context, params *DeleteBusinessMessagesParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "deleteBusinessMessages", params, &result)
+	return result, err
+}
+
+// SetBusinessAccountName https://core.telegram.org/bots/api#setbusinessaccountname
+func (b *Bot) SetBusinessAccountName(ctx context.Context, params *SetBusinessAccountNameParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "setBusinessAccountName", params, &result)
+	return result, err
+}
+
+// SetBusinessAccountUsername https://core.telegram.org/bots/api#setbusinessaccountusername
+func (b *Bot) SetBusinessAccountUsername(ctx context.Context, params *SetBusinessAccountUsernameParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "setBusinessAccountUsername", params, &result)
+	return result, err
+}
+
+// SetBusinessAccountBio https://core.telegram.org/bots/api#setbusinessaccountbio
+func (b *Bot) SetBusinessAccountBio(ctx context.Context, params *SetBusinessAccountBioParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "setBusinessAccountBio", params, &result)
+	return result, err
+}
+
+// SetBusinessAccountProfilePhoto https://core.telegram.org/bots/api#setbusinessaccountprofilephoto
+func (b *Bot) SetBusinessAccountProfilePhoto(ctx context.Context, params *SetBusinessAccountProfilePhotoParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "SetBusinessAccountProfilePhoto", params, &result)
+	return result, err
+}
+
+// RemoveBusinessAccountProfilePhoto https://core.telegram.org/bots/api#removebusinessaccountprofilephoto
+func (b *Bot) RemoveBusinessAccountProfilePhoto(ctx context.Context, params *RemoveBusinessAccountProfilePhotoParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "removeBusinessAccountProfilePhoto", params, &result)
+	return result, err
+}
+
+// SetBusinessAccountGiftSettings https://core.telegram.org/bots/api#setbusinessaccountgiftsettings
+func (b *Bot) SetBusinessAccountGiftSettings(ctx context.Context, params *SetBusinessAccountGiftSettingsParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "setBusinessAccountGiftSettings", params, &result)
+	return result, err
+}
+
+// GetBusinessAccountStarBalance https://core.telegram.org/bots/api#getbusinessaccountstarbalance
+func (b *Bot) GetBusinessAccountStarBalance(ctx context.Context, params *GetBusinessAccountStarBalanceParams) (*models.StarAmount, error) {
+	result := &models.StarAmount{}
+	err := b.rawRequest(ctx, "getBusinessAccountStarBalance", params, &result)
+	return result, err
+}
+
+// TransferBusinessAccountStars https://core.telegram.org/bots/api#transferbusinessaccountstars
+func (b *Bot) TransferBusinessAccountStars(ctx context.Context, params *TransferBusinessAccountStarsParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "transferBusinessAccountStars", params, &result)
+	return result, err
+}
+
+// GetBusinessAccountGifts https://core.telegram.org/bots/api#getbusinessaccountgifts
+func (b *Bot) GetBusinessAccountGifts(ctx context.Context, params *GetBusinessAccountGiftsParams) (*models.OwnedGifts, error) {
+	result := &models.OwnedGifts{}
+	err := b.rawRequest(ctx, "getBusinessAccountGifts", params, &result)
+	return result, err
+}
+
+// ConvertGiftToStars https://core.telegram.org/bots/api#convertgifttostars
+func (b *Bot) ConvertGiftToStars(ctx context.Context, params *ConvertGiftToStarsParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "convertGiftToStars", params, &result)
+	return result, err
+}
+
+// UpgradeGift https://core.telegram.org/bots/api#upgradegift
+func (b *Bot) UpgradeGift(ctx context.Context, params *UpgradeGiftParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "upgradeGift", params, &result)
+	return result, err
+}
+
+// TransferGift https://core.telegram.org/bots/api#transfergift
+func (b *Bot) TransferGift(ctx context.Context, params *TransferGiftParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "transferGift", params, &result)
+	return result, err
+}
+
+// PostStory https://core.telegram.org/bots/api#poststory
+func (b *Bot) PostStory(ctx context.Context, params *PostStoryParams) (*models.Story, error) {
+	result := &models.Story{}
+	err := b.rawRequest(ctx, "postStory", params, &result)
+	return result, err
+}
+
+// EditStory https://core.telegram.org/bots/api#editstory
+func (b *Bot) EditStory(ctx context.Context, params *EditStoryParams) (*models.Story, error) {
+	result := &models.Story{}
+	err := b.rawRequest(ctx, "editStory", params, &result)
+	return result, err
+}
+
+// DeleteStory https://core.telegram.org/bots/api#deletestory
+func (b *Bot) DeleteStory(ctx context.Context, params *DeleteStoryParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "deleteStory", params, &result)
+	return result, err
+}
+
+// GiftPremiumSubscription https://core.telegram.org/bots/api#giftpremiumsubscription
+func (b *Bot) GiftPremiumSubscription(ctx context.Context, params *GiftPremiumSubscriptionParams) (bool, error) {
+	var result bool
+	err := b.rawRequest(ctx, "giftPremiumSubscription", params, &result)
+	return result, err
+}