+++ models/business.go
@@ -2,12 +2,12 @@ package models
 
 // BusinessConnection https://core.telegram.org/bots/api#businessconnection
 type BusinessConnection struct {
-	ID         string `json:"id"`
-	User       User   `json:"user"`
-	UserChatID int64  `json:"user_chat_id"`
-	Date       int64  `json:"date"`
-	CanReply   bool   `json:"can_reply"`
-	IsEnabled  bool   `json:"is_enabled"`
+	ID         string             `json:"id"`
+	User       User               `json:"user"`
+	UserChatID int64              `json:"user_chat_id"`
+	Date       int64              `json:"date"`
+	Rights     *BusinessBotRights `json:"rights,omitempty"`
+	IsEnabled  bool               `json:"is_enabled"`
 }
 
 // BusinessMessagesDeleted https://core.telegram.org/bots/api#businessmessagesdeleted
@@ -40,4 +40,21 @@ type BusinessOpeningHours struct {
 type BusinessOpeningHoursInterval struct {
 	OpeningMinute int `json:"opening_minute"`
 	ClosingMinute int `json:"closing_minute"`
+}
+
+type BusinessBotRights struct {
+	CanReply                   bool `json:"can_reply,omitempty"`
+	CanReadMessages            bool `json:"can_read_messages,omitempty"`
+	CanDeleteOutgoingMessages  bool `json:"can_delete_outgoing_messages,omitempty"`
+	CanDeleteAllMessages       bool `json:"can_delete_all_messages,omitempty"`
+	CanEditName                bool `json:"can_edit_name,omitempty"`
+	CanEditBio                 bool `json:"can_edit_bio,omitempty"`
+	CanEditProfilePhoto        bool `json:"can_edit_profile_photo,omitempty"`
+	CanEditUsername            bool `json:"can_edit_username,omitempty"`
+	CanChangeGiftSettings      bool `json:"can_change_gift_settings,omitempty"`
+	CanViewGiftsAndStars       bool `json:"can_view_gifts_and_stars,omitempty"`
+	CanConvertGiftsToStars     bool `json:"can_convert_gifts_to_stars,omitempty"`
+	CanTransferAndUpgradeGifts bool `json:"can_transfer_and_upgrade_gifts,omitempty"`
+	CanTransferStars           bool `json:"can_transfer_stars,omitempty"`
+	CanManageStories           bool `json:"can_manage_stories,omitempty"`
 }